LayoutTests/imported/w3c/web-platform-tests/webcodecs/videoFrame-construction.any.worker-expected.txt


FAIL Test we can construct a VideoFrame. Can't find variable: OffscreenCanvas
FAIL Test closed VideoFrame. Can't find variable: OffscreenCanvas
FAIL Test we can construct a VideoFrame with a negative timestamp. Can't find variable: OffscreenCanvas
FAIL Test that timestamp is required when constructing VideoFrame from ImageBitmap promise_test: Unhandled rejection with value: object "ReferenceError: Can't find variable: OffscreenCanvas"
FAIL Test that timestamp is required when constructing VideoFrame from OffscreenCanvas promise_test: Unhandled rejection with value: object "ReferenceError: Can't find variable: OffscreenCanvas"
FAIL Test that timestamp is NOT required when constructing VideoFrame from another VideoFrame promise_test: Unhandled rejection with value: object "TypeError: Type error"
FAIL Test we can construct an odd-sized VideoFrame. Can't find variable: OffscreenCanvas
PASS Test constructing w/ unusable image argument throws: HAVE_NOTHING <video>.
FAIL Test constructing w/ unusable image argument throws: emtpy Canvas. Can't find variable: OffscreenCanvas
FAIL Test constructing w/ unusable image argument throws: closed ImageBitmap. Can't find variable: OffscreenCanvas
FAIL Test constructing w/ unusable image argument throws: closed VideoFrame. Can't find variable: OffscreenCanvas
FAIL Test invalid CanvasImageSource constructed VideoFrames Can't find variable: OffscreenCanvas
FAIL Test visibleRect metadata override where source display size = visible size Type error
FAIL Test visibleRect metadata override where source display width = 2 * visible width (anamorphic) Type error
FAIL Test visibleRect metadata override where source display size = 2 * visible size for both width and height Type error
FAIL Test visibleRect + display size metadata override Can't find variable: OffscreenCanvas
FAIL Test display size metadata override Can't find variable: OffscreenCanvas
FAIL Test invalid buffer constructed VideoFrames assert_throws_js: invalid coded size function "() => constructFrame({
                   timestamp: 1234,
                   codedWidth: Math.pow(2, 32) - 1,
                   codedHeight: Math.pow(2, 32) - 1,
                 })" did not throw
FAIL Test Uint8Array(ArrayBuffer) constructed I420 VideoFrame assert_equals: plane format expected (string) "I420" but got (object) null
FAIL Test ArrayBuffer constructed I420 VideoFrame assert_equals: plane format expected (string) "I420" but got (object) null
FAIL Test planar constructed I420 VideoFrame with colorSpace null is not an object (evaluating 'frame.colorSpace.primaries')
FAIL Test buffer constructed I420+Alpha VideoFrame assert_equals: plane format expected (string) "I420A" but got (object) null
FAIL Test buffer constructed NV12 VideoFrame assert_equals: plane format expected (string) "NV12" but got (object) null
FAIL Test buffer constructed RGB VideoFrames assert_equals: plane format expected (string) "RGBA" but got (object) null
FAIL Test VideoFrame constructed VideoFrame Can't find variable: OffscreenCanvas
FAIL Test we can construct a VideoFrame from an offscreen canvas. Can't find variable: OffscreenCanvas
FAIL Test I420 VideoFrame with odd visible size assert_equals: format expected (string) "I420" but got (object) null
FAIL Test I420A VideoFrame and alpha={keep,discard} assert_equals: plane format expected (string) "I420A" but got (object) null
FAIL Test RGBA, BGRA VideoFrames with alpha={keep,discard} assert_equals: plane format expected (string) "RGBA" but got (object) null
FAIL Test a VideoFrame constructed from canvas can drop the alpha channel. Can't find variable: OffscreenCanvas
